+++ src/pysm3/__init__.py
@@ -15,6 +15,7 @@ from .utils import (
     bandpass_unit_conversion,
     check_freq_input,
     set_verbosity,
+    apply_differential_smoothing,
     apply_smoothing_and_coord_transform,
     get_differential_beam_window,
     get_differential_fwhm,