+++ src/pysm3/utils/__init__.py
@@ -23,6 +23,7 @@ from .photometry import (
 )
 from .small_scales import sigmoid  # noqa: F401
 from .spherical_harmonics import (  # noqa: F401
+    apply_differential_smoothing,
     apply_smoothing_and_coord_transform,
     get_differential_beam_window,
     get_differential_fwhm,